requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langgraph>=0.2.0
pinecone>=5.0.0
openai>=1.0.0
streamlit>=1.40.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pytest>=8.0.0
